+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.0.5 - 2020-08-27
+Add condition type `ccobo.dateTimeNow`
+
+### Added
+- Condition type `ccobo.dateTimeNow`
+
 ## 0.0.4 - 2020-08-22
 Updated to generator-wp-dev-env#0.16.0 ( wp-dev-env-grunt#0.11.1 wp-dev-env-frame#0.11.0 )